ConstructionPreferences: Avoid raw asset references
Raw asset references are not tracked as dependencies and are prone to
breakage. Instead, use the game instance module to be able to reference
BP classes from C++ code.

And instead of hardcoding assumptions about what the conveyor belt
recipes are named, scan for them and build a list at startup. This
brings in the advantage of supporting modded belts as well.

Signed-off-by: Angel Pons <th3fanbus@gmail.com>